Connection pool implementation (#13)
* Add connection pool implementation

Key Features Implemented:

  - Connection pool with configurable size (MinConns, MaxConns)
  - Idle connection timeout (IdleTimeout)
  - Max connection lifetime (MaxLifetime)
  - Health checking - on acquire and periodic background checks
  - Pool statistics - gauges, counters, histograms, peak tracking
  - Wait timeout for exhausted pools
  - LIFO ordering for better cache locality
  - Waiter queue for blocking acquire when pool is exhausted
  - Thread-safe with proper locking and atomic operations
  - Observer pattern for metrics integration.

* Add metrics rate limit rate events and fix static bugs

+++ pkg/metrics/metrics_test.go
@@ -211,7 +211,7 @@ func TestCollectorConcurrency(t *testing.T) {
 		go func() {
 			for j := 0; j < 100; j++ {
 				c.SessionStarted()
-				c.RecordBytesSent(uint64(j))
+				c.RecordBytesSent(j)
 				c.RecordHandshakeLatency(time.Duration(j) * time.Millisecond)
 				c.SessionEnded()
 			}
